fixed a bug with spaces in the path of the sqlitefile
On linux systems the path of the sqlitefile (type nsIFile) has the
same format as an URI. But on Windows not. Therefore under windows the
xul-template in the options-dialog doesn't work. Also if there where
whitespaces in the path or filename the template wouldn't work
(windows and linux).
Now I create a nsIURI object from the nsIFile and so there is always a
correct uri for the template.